requirements.txt
web3==6.19.0
requests==2.32.3
psycopg[binary]==3.2.1
SQLAlchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
tenacity==8.5.0
streamlit==1.37.1
uvloop==0.19.0; sys_platform == 'linux'
pyyaml==6.0.2
psycopg2-binary==2.9.10